# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Gateway API - E-Commerce Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local (tanpa tunnel)"
        },
        {
            "url": "https://vegetarian-bloom-prospective-laid.trycloudflare.com/api",
            "description": "Local via Cloudflare Tunnel"
        }
    ],
    "paths": {
...
        "/payment/{orderId}": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create payment & get Midtrans Snap Token",
                "operationId": "a3b6e591075f239117496502220e4250",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Snap token created"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "400": {
                        "description": "Payment already processed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payment/notification": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Handle Midtrans notification webhook",
                "operationId": "a55c756db87eec4fcba8582e4f5d54c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "string"
                                    },
                                    "status_code": {
                                        "type": "string"
                                    },
                                    "gross_amount": {
                                        "type": "string"
                                    },
                                    "transaction_status": {
                                        "type": "string"
                                    },
                                    "fraud_status": {
                                        "type": "string"
                                    },
                                    "payment_type": {
                                        "type": "string"
                                    },
                                    "transaction_id": {
                                        "type": "string"
                                    },
                                    "signature_key": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification processed"
                    },
                    "403": {
                        "description": "Invalid signature"
                    }
                }
            }
        },
        "/payment/{orderId}/status": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Check payment status",
                "operationId": "96702720369dff8f9699b26cd76dd38b",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}